+++ cpp/src/toppra/geometric_path.hpp
@@ -15,6 +15,9 @@ namespace toppra {
  */
 class GeometricPath {
 public:
+
+  GeometricPath()  = default;
+
   /**
    * Constructor of GeometricPath on vector spaces.
    */
@@ -72,13 +75,16 @@ protected:
  */
 class PiecewisePolyPath : public GeometricPath {
 public:
+
+  PiecewisePolyPath() = default;
+
   /**
    * Consructor.
    *
    * @param coefficients Polynoamial coefficients.
    * @param breakpoints Vector of breakpoints.
    */
-  PiecewisePolyPath(const Matrices &, const std::vector<value_type> &);
+  PiecewisePolyPath(const Matrices&, std::vector<value_type>);
 
   /**
    * /brief Evaluate the path at given position.