+++ HISTORY.md
@@ -2,6 +2,12 @@
 
 ## 0.2.4 (in progress)
 
+## Added
+
+- #84, #85, #86, #87, #88: Initial development of the cpp codebase.
+
+## Changed
+
 - Simplify TOPPRA class interface.
 - Implement a new trajectory class for specified velocities.
 - [0e022c][cm-0e022c] Update README.md to reflect development roadmap.
@@ -9,6 +15,9 @@
 - [#78][gh-78] Improve documentation structure.
 - [#79][gh-79] Improve documentation structure.
 
+## Removed
+
+- Dropping support for Python2.7. Adding type annotation compatible to Python 3.
 
 ### 0.2.3 (Jan 25 2020)