+++ HISTORY.md
@@ -4,10 +4,11 @@
 
 ### Added
 
-- [cpp] Initial cpp TOPPRA implementation.
+- [cpp] Initial cpp TOPPRA implementation: compute parametrization, feasible sets.
 - [python] Implement a new trajectory class for specified velocities.
 
 ### Changed
+
 - [cpp] Bug fixes in solver wrapper.
 - [python] Simplify TOPPRA class interface.
 - [python] [0e022c][cm-0e022c] Update README.md to reflect development roadmap.
@@ -15,12 +16,15 @@
 - [python] [#78][gh-78] Improve documentation structure.
 - [python] [#79][gh-79] Improve documentation structure.
 
+### Removed
 
-### 0.2.3 (Jan 25 2020)
+- Dropping support for Python2.7. Adding type annotation compatible to Python 3.
+
+## 0.2.3 (Jan 25 2020)
 
 - Auto-gridpoints feature (#73)
 
-### <0.2.3
+## <0.2.3
 - Many commits and features.
 
 [gh-78]: https://github.com/hungpham2511/toppra/pull/78